refactor: 모집글 관련 form 기능 리팩토링
